move get_categorical combinations
